c/data-migrate: add get/set offsets RPC calls and router component
The RPC calls are to distribute/gather consumer offset data to/from
relevant partition leaders.

Router is a wrapper around leader routers for queries related to
consumer groups. This cannot belong to irpc_frontend because of circular
dependencies. It could go to frontend, but it's not its job: frontend
is a facade for migrations table and raft0, but this one operates
consumer group partitions.